Scene Change To  First Person Landscape works now!
Now once player lands successfully at a land site. The scene will change to the position on the land scape where the ship is placed and the ramp will open. The player can then navigate the terrain and go inside the ship. The ship is not interactable though. We could change this if we like but it's not necessary.