Game is nearing minimum viable product
The game functions nearly as intended. The player can land at landing sites and die. I want to add an exit button to the final scene so the player can exit the game whenever they would like at least maybe a minimap with the the location of the ship on it.